+++ biopipen/ns/scrna.py
@@ -1352,6 +1352,7 @@ class MarkersFinder(Proc):
             - plot_type: The type of the plot.
                 See <https://pwwang.github.io/biopipen.utils.R/reference/VizDEGs.html>.
                 Available types are `violin`, `box`, `bar`, `ridge`, `dim`, `heatmap` and `dot`.
+            - descr: A description of the plot to be shown above the plot image.
             - more_formats (type=list): The extra formats to save the plot in.
             - save_code (flag): Whether to save the code to generate the plot.
             - devpars (ns): The device parameters for the plots.
@@ -1363,6 +1364,7 @@ class MarkersFinder(Proc):
             The keys are the names of the cases and the values are the dicts inherited from `allmarker_plots_defaults`.
         allenrich_plots_defaults (ns): Default options for the plots to generate for the enrichment analysis.
             - plot_type: The type of the plot.
+            - descr: A description of the plot to be shown above the plot image.
             - devpars (ns): The device parameters for the plots.
                 - res (type=int): The resolution of the plots.
                 - height (type=int): The height of the plots.
@@ -1378,6 +1380,7 @@ class MarkersFinder(Proc):
                 See <https://pwwang.github.io/biopipen.utils.R/reference/VizDEGs.html>.
                 Available types are `violin`, `box`, `bar`, `ridge`, `dim`, `heatmap` and `dot`.
                 There are two additional types available - `volcano_pct` and `volcano_log2fc`.
+            - descr: A description of the plot to be shown above the plot image.
             - more_formats (type=list): The extra formats to save the plot in.
             - save_code (flag): Whether to save the code to generate the plot.
             - devpars (ns): The device parameters for the plots.
@@ -1394,6 +1397,7 @@ class MarkersFinder(Proc):
             - plot_type: The type of the plot.
                 See <https://pwwang.github.io/scplotter/reference/EnrichmentPlot.html>.
                 Available types are `bar`, `dot`, `lollipop`, `network`, `enrichmap` and `wordcloud`.
+            - descr: A description of the plot to be shown above the plot image.
             - more_formats (type=list): The extra formats to save the plot in.
             - save_code (flag): Whether to save the code to generate the plot.
             - devpars (ns): The device parameters for the plots.
@@ -1408,6 +1412,7 @@ class MarkersFinder(Proc):
             This means either `ident_1` should be empty, so that they can be expanded to multiple comparisons.
             - sigmarkers: The expression to filter the significant markers for each case.
                 If not provided, `envs.sigmarkers` will be used.
+            - descr: A description of the plot to be shown above the plot image.
             - plot_type (choice): The type of the plot to generate for the overlaps.
                 - venn: Use `plotthis::VennDiagram()`.
                 - upset: Use `plotthis::UpsetPlot()`.
@@ -1551,6 +1556,7 @@ class TopExpressingGenes(Proc):
             - plot_type: The type of the plot.
                 See <https://pwwang.github.io/scplotter/reference/EnrichmentPlot.html>.
                 Available types are `bar`, `dot`, `lollipop`, `network`, `enrichmap` and `wordcloud`.
+            - descr: A description of the plot to be shown above the plot image.
             - more_formats (type=list): The extra formats to save the plot in.
             - save_code (flag): Whether to save the code to generate the plot.
             - devpars (ns): The device parameters for the plots.
@@ -1982,6 +1988,7 @@ class ScFGSEA(Proc):
             See <https://rdrr.io/bioc/fgsea/man/fgseaMultilevel.html>
         alleach_plots_defaults (ns): Default options for the plots to generate for all pathways.
             - plot_type: The type of the plot, currently either dot or heatmap (default)
+            - descr: A description of the plot to be shown above the plot image.
             - devpars (ns): The device parameters for the plots.
                 - res (type=int): The resolution of the plots.
                 - height (type=int): The height of the plots.
@@ -3453,6 +3460,7 @@ class PseudoBulkDEG(Proc):
             - plot_type: The type of the plot.
                 See <https://pwwang.github.io/scplotter/reference/FeatureStatPlot.html>.
                 Available types are `violin`, `box`, `bar`, `ridge`, `dim`, `heatmap` and `dot`.
+            - descr: A description of the plot to be shown above the plot image.
             - more_formats (type=list): The extra formats to save the plot in.
             - save_code (flag): Whether to save the code to generate the plot.
             - devpars (ns): The device parameters for the plots.
@@ -3466,6 +3474,7 @@ class PseudoBulkDEG(Proc):
             The keys are the names of the cases and the values are the dicts inherited from `allmarker_plots_defaults`.
         allenrich_plots_defaults (ns): Default options for the plots to generate for the enrichment analysis.
             - plot_type: The type of the plot.
+            - descr: A description of the plot to be shown above the plot image.
             - devpars (ns): The device parameters for the plots.
                 - res (type=int): The resolution of the plots.
                 - height (type=int): The height of the plots.
@@ -3481,6 +3490,7 @@ class PseudoBulkDEG(Proc):
                 See <https://pwwang.github.io/scplotter/reference/FeatureStatPlot.html>.
                 Available types are `violin`, `box`, `bar`, `ridge`, `dim`, `heatmap` and `dot`.
                 There are two additional types available - `volcano_pct` and `volcano_log2fc`.
+            - descr: A description of the plot to be shown above the plot image.
             - more_formats (type=list): The extra formats to save the plot in.
             - save_code (flag): Whether to save the code to generate the plot.
             - devpars (ns): The device parameters for the plots.
@@ -3499,6 +3509,7 @@ class PseudoBulkDEG(Proc):
             - plot_type: The type of the plot.
                 See <https://pwwang.github.io/scplotter/reference/EnrichmentPlot.html>.
                 Available types are `bar`, `dot`, `lollipop`, `network`, `enrichmap` and `wordcloud`.
+            - descr: A description of the plot to be shown above the plot image.
             - more_formats (type=list): The extra formats to save the plot in.
             - save_code (flag): Whether to save the code to generate the plot.
             - devpars (ns): The device parameters for the plots.
@@ -3513,6 +3524,7 @@ class PseudoBulkDEG(Proc):
             This means either `ident_1` should be empty, so that they can be expanded to multiple comparisons.
             - sigmarkers: The expression to filter the significant markers for each case.
                 If not provided, `envs.sigmarkers` will be used.
+            - descr: A description of the plot to be shown above the plot image.
             - plot_type (choice): The type of the plot to generate for the overlaps.
                 - venn: Use `plotthis::VennDiagram()`.
                 - upset: Use `plotthis::UpsetPlot()`.